refactor : 보너스 번호와 우승 번호의 중복 코드를 제거해라
- LottoNumber와 관련된 검증 로직을 LottoNumberValidator에 두고, 사용하도록 변경
