+++ src/main/java/lotto/validator/LottoWinningNumbersValidator.java
@@ -11,12 +11,10 @@ public class LottoWinningNumbersValidator {
 
     private static final String LOTTO_WINNING_NUMBERS_PATTERN_TEXT = "^\\d+(,\\d+){5}$";
     private static final Pattern LOTTO_WINNING_NUMBERS_PATTERN = Pattern.compile(LOTTO_WINNING_NUMBERS_PATTERN_TEXT);
-    private final int lottoNumberMin;
-    private final int lottoNumberMax;
+    private final LottoNumberValidator lottoNumberValidator;
 
-    public LottoWinningNumbersValidator(int lottoNumberMin, int lottoNumberMax) {
-        this.lottoNumberMin = lottoNumberMin;
-        this.lottoNumberMax = lottoNumberMax;
+    public LottoWinningNumbersValidator(LottoNumberValidator lottoNumberValidator) {
+        this.lottoNumberValidator = lottoNumberValidator;
     }
 
     public void validateLottoWinningNumbers(String lottoWinningNumbers) {
@@ -27,25 +25,12 @@ public class LottoWinningNumbersValidator {
 
         winningNumbers.stream().forEach(lottoNumber -> {
             validateDuplicationNumbers(duplication, lottoNumber);
-            validateLessThanLottoNumberMax(lottoNumber);
-            validateMoreThanLottoNumberMin(lottoNumber);
+            lottoNumberValidator.validateLottoNumber(lottoNumber);
         });
     }
 
     private void validateFormat(String lottoWinningNumbers) {
         if (!LOTTO_WINNING_NUMBERS_PATTERN.matcher(lottoWinningNumbers).matches()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validateMoreThanLottoNumberMin(int lottoNumber) {
-        if (lottoNumber < lottoNumberMin) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validateLessThanLottoNumberMax(int lottoNumber) {
-        if (lottoNumber > lottoNumberMax) {
             throw new IllegalArgumentException();
         }
     }
